adding wmf alternatives 2
