+++ net/aclrian/messdiener/utils/ReadFile.java
@@ -13,7 +13,7 @@ import org.w3c.dom.NodeList;
 import net.aclrian.messdiener.deafault.Messdiener;
 import net.aclrian.messdiener.deafault.Messverhalten;
 import net.aclrian.messdiener.deafault.StandartMesse;
-import net.aclrian.messdiener.newy.progress.AData;
+import net.aclrian.messdiener.start.AData;
 
 /**
  * liesst aus einer xml-Datei den Messdiener und gibt diesen zurueck
