+++ net/aclrian/messdiener/window/auswaehlen/ATable.java
@@ -5,7 +5,7 @@ import javax.swing.table.DefaultTableModel;
 
 import net.aclrian.messdiener.deafault.Messverhalten;
 import net.aclrian.messdiener.deafault.StandartMesse;
-import net.aclrian.messdiener.newy.progress.AData;
+import net.aclrian.messdiener.start.AData;
 
 public class ATable extends JTable {
