+++ net/aclrian/messdiener/window/planerstellen/WMessenFrame.java
@@ -18,8 +18,8 @@ import com.toedter.calendar.JCalendar;
 import com.toedter.calendar.JDateChooser;
 import com.toedter.calendar.JTextFieldDateEditor;
 
-import net.aclrian.messdiener.newy.progress.AProgress;
 import net.aclrian.messdiener.pictures.References;
+import net.aclrian.messdiener.start.AProgress;
 import net.aclrian.messdiener.utils.Utilities;
 
 public class WMessenFrame extends JFrame {
